requirements.txt
python-dotenv
google-api-python-client==2.118.0
yt-dlp
faster-whisper==1.1.0
ctranslate2==4.6.0
nvidia-cublas-cu12
nvidia-cudnn-cu12==9.*
SQLAlchemy>=2.0